backend/data_scraper/requirements.txt
motor
httpx
python-dotenv